Ability to specify file download location
